Interfaz + boss
En la interfaz se muestra la oleada en la que estamos y los enemigos que quedan por derrotar para que esta oleada termine.
Se ha creado el boss. De momento solo se mueve y lanza el ataque a distancia. Queda implementar el resto de ataques.

+++ Alphas/Alphas/screen.h
@@ -22,8 +22,8 @@ protected:
 	EngineManager* m_engineManager;
 	ScreenManager* m_screenManager;
 
-	int	m_width;
-	int	m_height;
+	float	m_width;
+	float	m_height;
 
 	int		m_buttonFocused;
